Fix u-boot upgrade script
When having a Mender enabled bootloader and switching to a non-Mender-enabled
bootloader environment variables from the bootloader are set in a way that
the new bootloader will be written to the Mender default storage (that might
be the SD card). As the bootloader has to be stored to the internal eMMC
make sure, that it will be written there.

NGSL-244