Specify console in bootargs
If no console is specified in the bootargs, all available consoles will
be used. Thus we'll also have the console on the display. Whenever output
is generated on the display (e.g. kernel messages) this will update the
consoles, which will put the application to the background on the display.
With these settings in place, only the serial connection will be used as
console.

NGW-43